chore: extract modal logic to hooks
