+++ src/features/post/hooks/usePostCreationWizard.ts
@@ -1,0 +1,40 @@
+import { useState } from 'react';
+import { useSelector } from 'react-redux';
+import { RootState, useAppDispatch } from '@/shared/state/store';
+import { resetState } from '@/shared/state/slices/postSlice';
+import { closeModal } from '@/shared/state/slices/modalSlice';
+type UsePostCreationWizardResult = {
+  currentStep: string;
+  showExitConfirm: boolean;
+  handleCloseModal: () => void;
+  handleConfirmExit: () => void;
+  handleCancelExit: () => void;
+};
+
+export function usePostCreationWizard(): UsePostCreationWizardResult {
+  const { currentStep } = useSelector((state: RootState) => state.post);
+  const [showExitConfirm, setShowExitConfirm] = useState(false);
+  const dispatch = useAppDispatch();
+
+  const handleCloseModal = (): void => {
+    setShowExitConfirm(true);
+  };
+
+  const handleConfirmExit = (): void => {
+    dispatch(resetState());
+    dispatch(closeModal());
+    setShowExitConfirm(false);
+  };
+
+  const handleCancelExit = (): void => {
+    setShowExitConfirm(false);
+  };
+
+  return {
+    currentStep,
+    showExitConfirm,
+    handleCloseModal,
+    handleConfirmExit,
+    handleCancelExit,
+  };
+}
