+++ src/features/post/modal/AddPostSteps/CropStep/ThumbnailsPreview.tsx
